Split: keep only 2 file(s)
