+++ .gitignore
@@ -228,4 +228,3 @@ setup_leo.sh
 server.pid
 *.log
 package-lock.json
-llama.cpp/
