+++ studio/backend/core/export/export.py
@@ -8,28 +8,23 @@ Export backend - handles model exporting in various formats
 
 import glob
 import json
-import platform
 import structlog
 from loggers import get_logger
 import os
 import shutil
 from pathlib import Path
 from typing import Optional, Tuple, List
+from peft import PeftModel, PeftModelForCausalLM
 from unsloth import FastLanguageModel, FastVisionModel
 from huggingface_hub import HfApi, ModelCard
+from transformers.modeling_utils import PushToHubMixin
+import torch
 from utils.hardware import clear_gpu_cache
 
 from utils.models import is_vision_model, get_base_model_from_lora
 from utils.models.model_config import detect_audio_type
 from utils.paths import ensure_dir, outputs_root, resolve_export_dir, resolve_output_dir
 from core.inference import get_inference_backend
-
-# GPU-only imports — guarded for Apple Silicon where these aren't needed
-_IS_MLX = platform.system() == "Darwin" and platform.machine() == "arm64"
-if not _IS_MLX:
-    from peft import PeftModel, PeftModelForCausalLM
-    from transformers.modeling_utils import PushToHubMixin
-    import torch
 
 logger = get_logger(__name__)
 
@@ -228,7 +223,7 @@ class ExportBackend:
                 model, tokenizer = FastModel.from_pretrained(
                     model_name = checkpoint_path,
                     max_seq_length = max_seq_length,
-                    dtype = None if _IS_MLX else torch.float32,
+                    dtype = torch.float32,
                     load_in_4bit = False,
                     trust_remote_code = trust_remote_code,
                 )
@@ -265,12 +260,8 @@ class ExportBackend:
                     trust_remote_code = trust_remote_code,
                 )
 
-            # Check if PEFT / LoRA model
-            if _IS_MLX:
-                # MLX doesn't use PeftModel — detect LoRA via adapter_config.json
-                self.is_peft = adapter_config.exists()
-            else:
-                self.is_peft = isinstance(model, (PeftModel, PeftModelForCausalLM))
+            # Check if PEFT model
+            self.is_peft = isinstance(model, (PeftModel, PeftModelForCausalLM))
 
             # Store loaded model
             self.current_model = model
@@ -319,7 +310,7 @@ class ExportBackend:
         repo_id: Optional[str] = None,
         hf_token: Optional[str] = None,
         private: bool = False,
-    ) -> Tuple[bool, str]:
+    ) -> Tuple[bool, str, Optional[str]]:
         """
         Export merged model (for PEFT models).
 
@@ -332,7 +323,9 @@ class ExportBackend:
             private: Whether to make the repo private
 
         Returns:
-            Tuple of (success: bool, message: str)
+            Tuple of (success, message, output_path). output_path is the
+            resolved absolute on-disk directory of the saved model when
+            ``save_directory`` was set, else None.
         """
         if not self.current_model or not self.current_tokenizer:
             return False, "No model loaded. Please select a checkpoint first.", None
@@ -344,42 +337,31 @@ class ExportBackend:
                 None,
             )
 
+        output_path: Optional[str] = None
         try:
+            # Determine save method
+            if format_type == "4-bit (FP4)":
+                save_method = "merged_4bit_forced"
+            elif self._audio_type == "whisper":
+                # Whisper uses save_method=None for local 16-bit merged save
+                save_method = None
+            else:  # 16-bit (FP16)
+                save_method = "merged_16bit"
+
             # Save locally if requested
             if save_directory:
                 save_directory = str(resolve_export_dir(save_directory))
                 logger.info(f"Saving merged model locally to: {save_directory}")
                 ensure_dir(Path(save_directory))
 
-                if _IS_MLX:
-                    # Map UI's format_type to MLX save_method. LoRA and
-                    # GGUF go through separate export functions.
-                    mlx_save_method = (
-                        "merged_4bit"
-                        if format_type == "4-bit (FP4)"
-                        else "merged_16bit"
-                    )
-                    self.current_model.save_pretrained_merged(
-                        save_directory,
-                        self.current_tokenizer,
-                        save_method = mlx_save_method,
-                    )
-                else:
-                    # GPU: determine save method from format_type
-                    if format_type == "4-bit (FP4)":
-                        save_method = "merged_4bit_forced"
-                    elif self._audio_type == "whisper":
-                        save_method = None
-                    else:
-                        save_method = "merged_16bit"
-
-                    self.current_model.save_pretrained_merged(
-                        save_directory, self.current_tokenizer, save_method = save_method
-                    )
+                self.current_model.save_pretrained_merged(
+                    save_directory, self.current_tokenizer, save_method = save_method
+                )
 
                 # Write export metadata so the Chat page can identify the base model
                 self._write_export_metadata(save_directory)
                 logger.info(f"Model saved successfully to {save_directory}")
+                output_path = str(Path(save_directory).resolve())
 
             # Push to hub if requested
             if push_to_hub:
@@ -392,31 +374,20 @@ class ExportBackend:
 
                 logger.info(f"Pushing merged model to Hub: {repo_id}")
 
-                if _IS_MLX:
-                    # Reuse the local save_directory we just wrote to, so
-                    # push_to_hub_merged doesn't re-save under a relative
-                    # "username/model" directory in cwd.
-                    self.current_model.push_to_hub_merged(
-                        repo_id,
-                        self.current_tokenizer,
-                        save_directory = save_directory,
-                        token = hf_token,
-                        private = private,
-                    )
-                else:
-                    hub_save_method = (
-                        save_method if save_method is not None else "merged_16bit"
-                    )
-                    self.current_model.push_to_hub_merged(
-                        repo_id,
-                        self.current_tokenizer,
-                        save_method = hub_save_method,
-                        token = hf_token,
-                        private = private,
-                    )
+                # Whisper uses save_method=None for local but "merged_16bit" for hub push
+                hub_save_method = (
+                    save_method if save_method is not None else "merged_16bit"
+                )
+                self.current_model.push_to_hub_merged(
+                    repo_id,
+                    self.current_tokenizer,
+                    save_method = hub_save_method,
+                    token = hf_token,
+                    private = private,
+                )
                 logger.info(f"Model pushed successfully to {repo_id}")
 
-            return True, "Model exported successfully", None
+            return True, "Model exported successfully", output_path
 
         except Exception as e:
             logger.error(f"Error exporting merged model: {e}")
@@ -433,12 +404,14 @@ class ExportBackend:
         hf_token: Optional[str] = None,
         private: bool = False,
         base_model_id: Optional[str] = None,
-    ) -> Tuple[bool, str]:
+    ) -> Tuple[bool, str, Optional[str]]:
         """
         Export base model (for non-PEFT models).
 
         Returns:
-            Tuple of (success: bool, message: str)
+            Tuple of (success, message, output_path). output_path is the
+            resolved absolute on-disk directory of the saved model when
+            ``save_directory`` was set, else None.
         """
         if not self.current_model or not self.current_tokenizer:
             return False, "No model loaded. Please select a checkpoint first.", None
@@ -450,6 +423,7 @@ class ExportBackend:
                 None,
             )
 
+        output_path: Optional[str] = None
         try:
             # Save locally if requested
             if save_directory:
@@ -457,20 +431,13 @@ class ExportBackend:
                 logger.info(f"Saving base model locally to: {save_directory}")
                 ensure_dir(Path(save_directory))
 
-                if _IS_MLX:
-                    # MLX: save_pretrained_merged handles non-LoRA models too
-                    # (fuse() is a no-op when there are no LoRA layers)
-                    self.current_model.save_pretrained_merged(
-                        save_directory,
-                        self.current_tokenizer,
-                    )
-                else:
-                    self.current_model.save_pretrained(save_directory)
-                    self.current_tokenizer.save_pretrained(save_directory)
+                self.current_model.save_pretrained(save_directory)
+                self.current_tokenizer.save_pretrained(save_directory)
 
                 # Write export metadata so the Chat page can identify the base model
                 self._write_export_metadata(save_directory)
                 logger.info(f"Model saved successfully to {save_directory}")
+                output_path = str(Path(save_directory).resolve())
 
             # Push to hub if requested
             if push_to_hub:
@@ -483,63 +450,46 @@ class ExportBackend:
 
                 logger.info(f"Pushing base model to Hub: {repo_id}")
 
-                if _IS_MLX:
-                    # Reuse the local save_directory to avoid a redundant
-                    # re-save under a relative "username/model" path.
-                    self.current_model.push_to_hub_merged(
-                        repo_id,
-                        self.current_tokenizer,
-                        save_directory = save_directory,
-                        token = hf_token,
-                        private = private,
+                # Get base model name from request or model config
+                base_model = (
+                    base_model_id
+                    or self.current_model.config._name_or_path
+                    or "unknown"
+                )
+
+                # Create repo
+                hf_api = HfApi(token = hf_token)
+                repo_id = PushToHubMixin._create_repo(
+                    PushToHubMixin,
+                    repo_id = repo_id,
+                    private = private,
+                    token = hf_token,
+                )
+                username = repo_id.split("/")[0]
+
+                # Create and push model card
+                content = MODEL_CARD.format(
+                    username = username,
+                    base_model = base_model,
+                    model_type = self.current_model.config.model_type,
+                    method = "",
+                    extra = "unsloth",
+                )
+                card = ModelCard(content)
+                card.push_to_hub(
+                    repo_id, token = hf_token, commit_message = "Unsloth Model Card"
+                )
+
+                # Upload model files
+                if save_directory:
+                    hf_api.upload_folder(
+                        folder_path = save_directory, repo_id = repo_id, repo_type = "model"
                     )
+                    logger.info(f"Model pushed successfully to {repo_id}")
                 else:
-                    # Get base model name from request or model config
-                    base_model = (
-                        base_model_id
-                        or self.current_model.config._name_or_path
-                        or "unknown"
-                    )
+                    return False, "Local save directory required for Hub upload", None
 
-                    # Create repo
-                    hf_api = HfApi(token = hf_token)
-                    repo_id = PushToHubMixin._create_repo(
-                        PushToHubMixin,
-                        repo_id = repo_id,
-                        private = private,
-                        token = hf_token,
-                    )
-                    username = repo_id.split("/")[0]
-
-                    # Create and push model card
-                    content = MODEL_CARD.format(
-                        username = username,
-                        base_model = base_model,
-                        model_type = self.current_model.config.model_type,
-                        method = "",
-                        extra = "unsloth",
-                    )
-                    card = ModelCard(content)
-                    card.push_to_hub(
-                        repo_id, token = hf_token, commit_message = "Unsloth Model Card"
-                    )
-
-                    # Upload model files
-                    if save_directory:
-                        hf_api.upload_folder(
-                            folder_path = save_directory,
-                            repo_id = repo_id,
-                            repo_type = "model",
-                        )
-                        logger.info(f"Model pushed successfully to {repo_id}")
-                    else:
-                        return (
-                            False,
-                            "Local save directory required for Hub upload",
-                            None,
-                        )
-
-            return True, "Model exported successfully", None
+            return True, "Model exported successfully", output_path
 
         except Exception as e:
             logger.error(f"Error exporting base model: {e}")
@@ -555,7 +505,7 @@ class ExportBackend:
         push_to_hub: bool = False,
         repo_id: Optional[str] = None,
         hf_token: Optional[str] = None,
-    ) -> Tuple[bool, str]:
+    ) -> Tuple[bool, str, Optional[str]]:
         """
         Export model in GGUF format.
 
@@ -567,11 +517,14 @@ class ExportBackend:
             hf_token: Hugging Face token
 
         Returns:
-            Tuple of (success: bool, message: str)
+            Tuple of (success, message, output_path). output_path is the
+            resolved absolute on-disk directory containing the .gguf
+            files when ``save_directory`` was set, else None.
         """
         if not self.current_model or not self.current_tokenizer:
             return False, "No model loaded. Please select a checkpoint first.", None
 
+        output_path: Optional[str] = None
         try:
             # Convert quantization method to lowercase for unsloth
             quant_method = quantization_method.lower()
@@ -666,6 +619,7 @@ class ExportBackend:
                     abs_save_dir,
                     "\n  ".join(os.path.basename(f) for f in final_ggufs) or "(none)",
                 )
+                output_path = str(Path(abs_save_dir).resolve())
 
             # Push to hub if requested
             if push_to_hub:
@@ -689,7 +643,7 @@ class ExportBackend:
             return (
                 True,
                 f"GGUF model exported successfully ({quantization_method})",
-                None,
+                output_path,
             )
 
         except Exception as e:
@@ -706,12 +660,14 @@ class ExportBackend:
         repo_id: Optional[str] = None,
         hf_token: Optional[str] = None,
         private: bool = False,
-    ) -> Tuple[bool, str]:
+    ) -> Tuple[bool, str, Optional[str]]:
         """
         Export LoRA adapter only (not merged).
 
         Returns:
-            Tuple of (success: bool, message: str)
+            Tuple of (success, message, output_path). output_path is the
+            resolved absolute on-disk directory of the saved adapter
+            when ``save_directory`` was set, else None.
         """
         if not self.current_model or not self.current_tokenizer:
             return False, "No model loaded. Please select a checkpoint first.", None
@@ -719,6 +675,7 @@ class ExportBackend:
         if not self.is_peft:
             return False, "This is not a PEFT model. No adapter to export.", None
 
+        output_path: Optional[str] = None
         try:
             # Save locally if requested
             if save_directory:
@@ -726,14 +683,10 @@ class ExportBackend:
                 logger.info(f"Saving LoRA adapter locally to: {save_directory}")
                 ensure_dir(Path(save_directory))
 
-                if _IS_MLX:
-                    # MLX: save adapters.safetensors + tokenizer files
-                    self.current_model.save_lora_adapters(save_directory)
-                    self.current_tokenizer.save_pretrained(save_directory)
-                else:
-                    self.current_model.save_pretrained(save_directory)
-                    self.current_tokenizer.save_pretrained(save_directory)
+                self.current_model.save_pretrained(save_directory)
+                self.current_tokenizer.save_pretrained(save_directory)
                 logger.info(f"Adapter saved successfully to {save_directory}")
+                output_path = str(Path(save_directory).resolve())
 
             # Push to hub if requested
             if push_to_hub:
@@ -746,30 +699,13 @@ class ExportBackend:
 
                 logger.info(f"Pushing LoRA adapter to Hub: {repo_id}")
 
-                if _IS_MLX:
-                    # Save locally first, then upload folder
-                    import tempfile
-
-                    with tempfile.TemporaryDirectory() as tmp_dir:
-                        self.current_model.save_lora_adapters(tmp_dir)
-                        self.current_tokenizer.save_pretrained(tmp_dir)
-                        hf_api = HfApi(token = hf_token)
-                        hf_api.create_repo(repo_id, private = private, exist_ok = True)
-                        hf_api.upload_folder(
-                            folder_path = tmp_dir,
-                            repo_id = repo_id,
-                            repo_type = "model",
-                        )
-                else:
-                    self.current_model.push_to_hub(
-                        repo_id, token = hf_token, private = private
-                    )
-                    self.current_tokenizer.push_to_hub(
-                        repo_id, token = hf_token, private = private
-                    )
+                self.current_model.push_to_hub(repo_id, token = hf_token, private = private)
+                self.current_tokenizer.push_to_hub(
+                    repo_id, token = hf_token, private = private
+                )
                 logger.info(f"Adapter pushed successfully to {repo_id}")
 
-            return True, "LoRA adapter exported successfully", None
+            return True, "LoRA adapter exported successfully", output_path
 
         except Exception as e:
             logger.error(f"Error exporting LoRA adapter: {e}")
