+++ studio/frontend/src/config/env.ts
@@ -50,7 +50,7 @@ export async function fetchDeviceType(): Promise<DeviceType> {
     if (res.ok) {
       const data = (await res.json()) as { device_type?: string; chat_only?: boolean };
       const deviceType = data.device_type ?? detectLocalPlatform();
-      const chatOnly = data.chat_only ?? false;
+      const chatOnly = data.chat_only ?? deviceType === "mac";
       usePlatformStore.setState({ deviceType, chatOnly, fetched: true });
       return deviceType;
     }
