+++ studio/frontend/src/features/studio/sections/params-section.tsx
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 // Copyright 2026-present the Unsloth AI Inc. team. All rights reserved. See /studio/LICENSE.AGPL-3.0
 
-import { usePlatformStore } from "@/config/env";
 import { SectionCard } from "@/components/section-card";
 import { Checkbox } from "@/components/ui/checkbox";
 import {
@@ -884,11 +883,7 @@ export function ParamsSection(): ReactElement {
                       <SelectContent>
                         <SelectItem value="none">None</SelectItem>
                         <SelectItem value="true">Standard</SelectItem>
-                        {usePlatformStore.getState().deviceType === "mac" ? (
-                          <SelectItem value="mlx">MLX</SelectItem>
-                        ) : (
-                          <SelectItem value="unsloth">Unsloth</SelectItem>
-                        )}
+                        <SelectItem value="unsloth">Unsloth</SelectItem>
                       </SelectContent>
                     </Select>
                   </Row>
